backend/requirements.txt
# AI-Powered Code Modernization Platform
# Python Dependencies

# Core AI & Agent Framework
anthropic>=0.18.0
openai>=1.35.0
google-generativeai>=0.8.0
huggingface_hub>=0.20.0
dashscope>=1.13.0
mcp>=0.1.0
langgraph>=0.2.16
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.0.1

# Utilities
python-dotenv>=1.0.0

# Backend API
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Docker & Container Management
docker>=6.1.0

# GitHub Integration
pygithub>=2.1.0

# Database & State Management
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Code Analysis
# ast-grep is a CLI tool, not a Python package - install separately if needed
# tree-sitter is optional for advanced code parsing
# tree-sitter>=0.20.0

# Logging & Monitoring
structlog>=23.2.0
rich>=13.7.0

# Cost Tracking
tiktoken>=0.5.0
